backend/requirements.txt
fastapi
uvicorn
yfinance
ddgs
anthropic
python-dotenv
pydantic
